workspace/requirements.txt
chainlit>=2.0.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
boto3>=1.35.0
psycopg2-binary>=2.9.9
bitwarden-sdk>=2.0.0
